Introduced new Sink object for parsing
